requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
requests
httpx
pytest
discord.py
websockets
